Added server and updated dashboard/ui.
The server code in it's clean state added. The history is missing for it
as it contained private information.

The dashboard ui has had the websocket url moved to an easier place for
editing.

Tagged the phone number in the car-thing.html to make it easier to find.

Added documentation to the dashboard in the form of a README on how to
get it running.